Move drop_privileges to capabilities.c
